fix bug action type in dqn

+++ rl_bot/value_based/ddqn.py
@@ -102,12 +102,12 @@ class DDQN:
         with torch.no_grad():
             actions = torch.argmax(self.policy_net(data["next_states"]), dim=1)
             # calculate q_target
-            q_targets = torch.gather(self.target_net(data["next_states"]), 1, actions.unsqueeze(1)).squeeze(1)
+            q_targets = torch.gather(self.target_net(data["next_states"]), 1, actions.unsqueeze(1).long()).squeeze(1)
             # y_j = r_j if s_j+1 is terminal state
             # else r_j + gamma*q_targets
             target = data["rewards"] + self.args["gamma"] * q_targets * (1 - data["dones"])
 
-        q_values = torch.gather(self.policy_net(data["states"]), 1, data["actions"].unsqueeze(1)).squeeze(1)
+        q_values = torch.gather(self.policy_net(data["states"]), 1, data["actions"].unsqueeze(1).long()).squeeze(1)
 
         # typical backward pass
         loss = F.mse_loss(target, q_values)
